Box2dPhysics now works with convex polygons. Fixed bug in Camera class. Demo4-physics has been extended.

+++ Dodge/include/dodge/math/primitives/Ellipse.hpp
@@ -23,7 +23,7 @@ class Ellipse : public Primitive {
 
       virtual Vec2f getMinimum() const;
       virtual Vec2f getMaximum() const;
-      virtual void rotate(double rads, const Vec2f& pivot);
+      virtual void rotate(float32_t rads, const Vec2f& pivot);
       virtual void scale(const Vec2f& sv);
 #ifdef DEBUG
       virtual void dbg_print(std::ostream& out, int tab) const;
